+++ vulkano/src/pipeline/vertex/mod.rs
@@ -76,6 +76,7 @@ pub use self::vertex::Vertex;
 pub use self::vertex::VertexMemberInfo;
 pub use self::vertex::VertexMemberTy;
 pub use self::bufferless::BufferlessDefinition;
+pub use self::bufferless::BufferlessVertices;
 
 mod definition;
 mod impl_vertex;