+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Changed `CpuBufferPool::next()` and `chunk()` to return a `Result` in case of an error when
   allocating or mapping memory.
 - Fixed `layers` argument validation in `Swapchain::new_inner`.
+- Added `vulkano::pipeline::vertex::BufferlessDefinition` and `BufferlessVertices` to enable
+  bufferless drawing, which is useful when drawing with shaders that compute their vertices
+  internally.
 
 # Version 0.6.2 (2017-09-06)